Initial commit of a threaded implementation of OpenCat.
This is not yet working. Right now the ConfigureNybble (replacement
for WriteInstinct) will compile, but does not operate properly. The next
step is to clean up the servo control so that it actually uses the
protothreading approach instead of making blocking calls. Ultimately the
ThreadedNybble code (until it gets a better name) should be made as a library
with two example ino files - one for Configuration and on for normal running.